+++ tests/scenario/test_area.py
@@ -50,10 +50,18 @@ class TestArea(unittest.TestCase):
         self.assertEqual(area_border.boundary, 2)
 
         # test the properties of adjacent
-        area_border.adjacent = 1
+        area_border.adjacent = [1]
         with self.assertRaises(AssertionError):
             area_border.adjacent = "a"
-        self.assertEqual(area_border.adjacent, 1)
+        self.assertEqual(area_border.adjacent, [1])
+        area_border.adjacent = [1, 2, 3]
+        with self.assertRaises(AssertionError):
+            area_border.adjacent = [1, "a", 3]
+        with self.assertRaises(AssertionError):
+            area_border.adjacent = 1
+        with self.assertRaises(AssertionError):
+            area_border.adjacent = ["a", "b", "c"]
+        self.assertEqual(area_border.adjacent, [1, 2, 3])
 
         # test the properties of line_marking
         area_border.line_marking = LineMarking.DASHED