+++ tutorials/code_example.py
@@ -1,6 +1,4 @@
 # import functions to read xml file and visualize commonroad objects
-import matplotlib.pyplot as plt
-
 from commonroad.common.file_reader import CommonRoadFileReader
 from commonroad.visualization.mp_renderer import MPRenderer
 
@@ -10,11 +8,7 @@ file_path = "ZAM_Tutorial-1_1_T-1.pb"
 # read in the scenario and planning problem set
 scenario, planning_problem_set, _ = CommonRoadFileReader(filename_dynamic=file_path).open_all()
 
-# plot the planning problem and the scenario for the fifth time step
-plt.figure(figsize=(25, 10))
-rnd = MPRenderer()
-rnd.draw_params.time_begin = 5
+# plot the scenario
+rnd = MPRenderer(figsize=(25, 10))
 scenario.draw(rnd)
-planning_problem_set.draw(rnd)
-rnd.render()
-plt.show()
+rnd.render(show=True)